# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a sample server Petstore server.  You can find out more about Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/).  For this sample, you can use the api key `special-key` to test the authorization filters.",
    "version": "1.0.0",
    "title": "T-Coffee API"
  },
  "host": "localhost:3000",
  "basePath": "/api/v1",
  "tags": [
    {
      "name": "products",
      "description": "Everything about your Products"
    }
  ],
  "schemes": ["https", "http"],
  "paths": {
    "/product": {
      "get": {
        "tags": ["product"],
        "summary": "Get all products",
        "description": "Returns all products from the system that the user has access to",
        "operationId": "getProducts",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Product"
              }
            }
          },
          "400": {
            "description": "Invalid status value"
          }
        }
      },
      "post": {
        "tags": ["product"],
        "summary": "Add a new product",
        "description": "Adds a new product to the store",
        "operationId": "addProduct",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Product object that needs to be added to the store",
            "required": true,
            "schema": {
              "$ref": "#/definitions/productCreate"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/ProductCreateResponse"
            }
          },
          "409": {
            "description": "Product already exists"
          },
          "404": {
            "description": "Invalid input"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/products/{productId}": {
      "get": {
        "tags": ["product"],
        "summary": "Get a product by ID",
        "description": "",
        "operationId": "getProductById",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "description": "ID of product to return",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Product"
            }
          },
          "500": {
            "description": "Invalid ID supplied"
          },
          "400": {
            "description": "Please provide product id"
          },
          "404": {
            "description": "Product not found"
          }
        }
      },
      "put": {
        "tags": ["product"],
        "summary": "Update an existing product",
        "description": "",
        "operationId": "updateProduct",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "description": "ID of product that needs to be updated",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Product object that needs to be updated",
            "schema": {
              "$ref": "#/definitions/updateProduct"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Product updated successfully"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Product not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": ["product"],
        "summary": "Delete a product",
        "description": "",
        "operationId": "deleteProduct",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "description": "ID of product that needs to be deleted",
            "required": true,
            "type": "string",
            "format": "uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Product deleted successfully"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Product not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
  "definitions": {
    "Product": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "uuid"
        },
        "name": {
          "type": "string"
        },
        "price": {
          "type": "number",
          "format": "double"
        },
        "description": {
          "type": "string"
        },
        "image": {
          "type": "string"
        },
        "variants": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Variant"
          }
        },
        "categories": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Category"
          }
        }
      }
    },
    "Variant": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "type": {
          "type": "string"
        },
        "options": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Option"
          }
        }
      }
    },
    "Option": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "additional_price": {
          "type": "number",
          "format": "double"
        },
        "variant_id": {
          "type": "string",
          "format": "uuid"
        }
      }
    },
    "Category": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "icon": {
          "type": "string"
        }
      }
    },

    "productCreate": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "Cappuccino"
        },
        "price": {
          "type": "number",
          "format": "double",
          "example": 39000
        },
        "description": {
          "type": "string",
          "example": "Cappuccino is a coffee drink that today is typically composed of double espresso and hot milk, with the surface topped with foamed milk."
        },
        "image": {
          "type": "string",
          "example": "https://www.starbucks.com.vn/media/catalog/product/cache/1/image/9df78eab33525d08d6e5fb8d27136e95/c/a/cappuccino.jpg"
        },
        "variantId": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "example": ["size", "sugar", "topping"]
        },
        "categoryId": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "example": ["coffee", "drinks"]
        }
      }
    },

    "ProductCreateResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "example": "Product created successfully"
        },
        "result": {
          "type": "object",
          "properties": {
            "id": {
              "type": "string",
              "format": "uuid",
              "example": "3a798e4b-fee4-4f0b-b4f8-7e7f2af67e1e"
            },
            "name": {
              "type": "string",
              "example": "Hot Capuchino"
            },
            "image": {
              "type": "string",
              "example": "https://stc-zmp.zadn.vn/templates/zaui-coffee/dummy/product-square-1.webp"
            }
          }
        }
      }
    },
    "updateProduct": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "Cappuccino",
          "required": ["false"]
        },
        "price": {
          "type": "number",
          "format": "double",
          "example": 39000,
          "required": ["false"]
        },
        "description": {
          "type": "string",
          "example": "Cappuccino is a coffee drink that today is typically composed of double espresso and hot milk, with the surface topped with foamed milk.",
          "required": ["false"]
        },
        "image": {
          "type": "string",
          "example": "https://www.starbucks.com.vn/media/catalog/product/cache/1/image/9df78eab33525d08d6e5fb8d27136e95/c/a/cappuccino.jpg",
          "required": ["false"]
        },
        "variantId": {
          "type": "array",
          "required": ["false"],
          "items": {
            "type": "string"
          },
          "example": ["size", "sugar", "topping"]
        },
        "categoryId": {
          "type": "array",
          "required": ["false"],
          "items": {
            "type": "string"
          },
          "example": ["coffee", "drinks"]
        }
      }
    }
  }
}